+++ .gitignore
@@ -32,3 +32,5 @@ fastn-core/tests/**/.packages/fifthtry.github.io/package-info/
 docs
 .env
 .env.swp
+
+fastn/gen
