+++ Cargo.toml
@@ -112,6 +112,7 @@ include_dir = "0.7"
 indexmap = { version = "2", features = ["serde"] }
 indoc = "2"
 itertools = "0.14"
+log = "0.4"
 mime_guess = "2"
 once_cell = "1"
 prettify-js = "0.1.0"
@@ -133,6 +134,9 @@ thiserror = "2"
 tokio = { version = "1", features = ["full"] }
 tokio-postgres = { version = "0.7", features = ["with-serde_json-1", "with-uuid-1"] }
 tracing = "0.1"
+tauri = "2.6"
+tauri-build = { version = "2", features = ["config-json5"] }
+tauri-plugin-log = "2"
 url = "2"
 walkdir = "2"
 wasmtime = "34"
