+++ fastn/src/tauri.rs
@@ -1,0 +1,18 @@
+#[allow(unexpected_cfgs)]
+#[cfg_attr(mobile, tauri::mobile_entry_point)]
+pub async fn run() {
+  tauri::Builder::default()
+    .setup(|app| {
+      if cfg!(debug_assertions) {
+        app.handle().plugin(
+          tauri_plugin_log::Builder::default()
+            .level(log::LevelFilter::Info)
+            .build(),
+        )?;
+      }
+      Ok(())
+    })
+    .run(tauri::generate_context!())
+    .expect("error while running tauri application");
+}
+
