+++ fastn/web_assets/index.html
@@ -1,0 +1,1 @@
+<h1>Hello fastn!</h1>
