+++ flake.nix
@@ -1,5 +1,6 @@
 {
   inputs = {
+    nixpkgs.url = "github:nixos/nixpkgs/nixos-unstable";
     flake-utils.url = "github:numtide/flake-utils";
     rust-overlay = {
       url = "github:oxalica/rust-overlay";
@@ -28,6 +29,8 @@
             openssl.dev
             diesel-cli
             rust-analyzer-unwrapped
+
+            cargo-tauri
           ] ++ lib.optionals stdenv.isDarwin [ darwin.apple_sdk.frameworks.Foundation ];
 
           shellHook = ''
